back/requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
fastapi-mail==1.4.1
fastapi-sso==0.17.0
fastapi-storages==0.3.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.6
oauthlib==3.2.2
orjson==3.10.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.6.1
pydantic_core==2.18.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.2.0
requests==2.31.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.29
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0